Move utility functions to dedicated space
This moves all the utility files to a dedicated folder.
